+++ Changelog.md
@@ -4,6 +4,8 @@ OpenCore Changelog
 - Improved SSDT-PNLF compatibility with CFL+ graphics
 - Fixed OpenCanopy performance loss due to redrawing introduced in 0.6.9
 - Added pattern-based automatic variable initialisation for better security
+- Updated underlying EDK II package to edk2-stable202108
+- Updated Apple Secure Boot variables for `x86legacy`
 
 #### v0.7.2
 - Fixed OSBundleLibraries/OSBundleLibaries64 handling
@@ -15,7 +17,6 @@ OpenCore Changelog
 - Updated builtin firmware versions for SMBIOS and the rest
 - Improved SSDT-PNLF compatibility with Windows and newer graphics
 - Fixed CLANGPDB OpenCore builds by shortening OC magic
-- Updated underlying EDK II package to edk2-stable202108
 
 #### v0.7.1
 - Added `SyncTableIds` quirk to sync modified table OEM identifiers